third_party/WebKit/LayoutTests/fullscreen/full-screen-request-disallow-for-non-media-elements-expected.txt

EVENT(webkitfullscreenerror)
EVENT(webkitfullscreenchange)
END OF TEST
Tests for the disallowFullscreenForNonMediaElements setting which will be set in those platforms where fullscreen for non-media elements is not supported.
